refactor: move cli components from api to cli sub-package
