+++ junifer/api/__init__.py
@@ -5,8 +5,6 @@
 # License: AGPL
 
 from . import decorators
-from .cli import cli
-from .functions import collect, queue, run
 
 
-__all__ = ["decorators", "cli", "collect", "queue", "run"]
+__all__ = ["decorators"]
