+++ junifer/cli/__init__.py
@@ -1,0 +1,10 @@
+"""Junifer CLI."""
+
+# Authors: Synchon Mandal <s.mandal@fz-juelich.de>
+# License: AGPL
+
+from .cli import cli
+from .functions import collect, queue, run
+
+
+__all__ = ["cli", "collect", "queue", "run", "reset", "list_elements"]
