+++ junifer/cli/tests/data/gmd_mean.yaml
