+++ junifer/cli/tests/data/partly_cloudy_agg_mean_tian.yml
