+++ junifer/cli/tests/test_cli_utils.py
@@ -9,7 +9,7 @@ import sys
 import pytest
 
 from junifer._version import __version__
-from junifer.api.utils import (
+from junifer.cli.utils import (
     _get_dependency_information,
     _get_environment_information,
     _get_junifer_version,
