+++ pyproject.toml
@@ -62,7 +62,7 @@ Homepage = "https://juaml.github.io/junifer"
 Source = "https://github.com/juaml/junifer"
 
 [project.scripts]
-junifer = "junifer.api.cli:cli"
+junifer = "junifer.cli.cli:cli"
 
 [project.optional-dependencies]
 all = [
